+++ createpythonpackage/templates/test_project.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-import unittest
-
-
-class UnitTests(unittest.TestCase):
-    def test_project(self):
-        raise NotImplementedError("Add your tests here to make them pass")